minor navigation edit to place the btn size login in the menu button jsx element
