+++ src/Components/Layout/Navigation.js
@@ -1,20 +1,11 @@
 import React, { useEffect, useState } from "react";
 import MobileNavButton from "../UI/MobileNavButton";
-import { useMediaQuery } from "../../hooks/useMediaQuery";
 import "../../css/Navigation.css";
 
 
 const Navigation = () => {
 
-
-    const isLargeScreen = useMediaQuery("lg");
-    const isMediumScreen = useMediaQuery("md");
-    const isSmallScreen = useMediaQuery("sm");
-    const isExtraSmallScreen = useMediaQuery("xs");
-
-
     const [isMobileNavOpen, setIsMobileNavOpen] = useState(false);
-    const [mobileNavBtnSize, setMobileNavBtnSize] = useState("");
 
 
     const handleMobileNavClick = () => {
@@ -23,25 +14,13 @@ const Navigation = () => {
 
 
     useEffect(() => {
-        if (isLargeScreen) {
-            setMobileNavBtnSize("50px");
-        } else if (isMediumScreen) {
-            setMobileNavBtnSize("40px");
-        } else if (isSmallScreen) {
-            setMobileNavBtnSize("40px");
-        } else if (isExtraSmallScreen) {
-            setMobileNavBtnSize("30px");
-        }
-    }, [isLargeScreen, isMediumScreen, isSmallScreen, isExtraSmallScreen]);
-
-    useEffect(() => {
         const sectionAboutEl = document.querySelector("#about");
         const observer = new IntersectionObserver((entries) => {
             const ent = entries[0];
-            if(!ent.isIntersecting) {
+            if (!ent.isIntersecting) {
                 document.body.classList.add("sticky");
             }
-            if(ent.isIntersecting) {
+            if (ent.isIntersecting) {
                 document.body.classList.remove("sticky");
             }
         }, { root: null, threshold: 0 })
@@ -49,7 +28,7 @@ const Navigation = () => {
     }, []);
 
 
-    
+
     return (
 
         <header className={`header ${isMobileNavOpen ? "nav-open" : ""}  `}>
@@ -68,7 +47,7 @@ const Navigation = () => {
                 </ul>
             </nav>
 
-            <MobileNavButton handleMobileNavClick={handleMobileNavClick} mobileNavBtnSize={mobileNavBtnSize} />
+            <MobileNavButton handleMobileNavClick={handleMobileNavClick} />
 
         </header>
     );
